调整migrate序号 #3099
