+++ src/scene_server/admin_server/imports.go
@@ -50,6 +50,6 @@ import (
 	_ "configcenter/src/scene_server/admin_server/upgrader/x19.08.20.01"
 	_ "configcenter/src/scene_server/admin_server/upgrader/x19.08.26.01"
 	_ "configcenter/src/scene_server/admin_server/upgrader/x19.08.26.02"
-	_ "configcenter/src/scene_server/admin_server/upgrader/x19.08.26.03"
 	_ "configcenter/src/scene_server/admin_server/upgrader/x19.09.03.01"
+	_ "configcenter/src/scene_server/admin_server/upgrader/x19.09.03.02"
 )
